+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Forward kinematics and inverse kinematics of `four_dof_scara_robot`([!38](http://101.42.250.169:8081/robodyno/robodyno/-/merge_requests/38)).
+- README for a `four_dof_scara_robot`([!38](http://101.42.250.169:8081/robodyno/robodyno/-/merge_requests/38)).
+
 ### Added
 - `ProximitySensor` component([!36](http://101.42.250.169:8081/robodyno/robodyno/-/merge_requests/36)).
 - Examples of ProximitySensor([!36](http://101.42.250.169:8081/robodyno/robodyno/-/merge_requests/36)).